+++ CHANGELOG.md
@@ -14,7 +14,6 @@ Types of changes:
 
 ## 4.0.0
 ### Changed
-- `<Autocomplete>` - New Autocomplete component from wix-ui-backoffice [1620](https://github.com/wix/wix-style-react/pull/1620)
 - `<ToggleSwitch>` - New ToggleSwitch component from wix-ui-backoffice [1637](https://github.com/wix/wix-style-react/pull/1637)
