--- src/Deprecated/AutoComplete/AutoComplete.driver.js
@@ -1,5 +1,0 @@
-import inputWithOptionsDriverFactory from '../../InputWithOptions/InputWithOptions.driver';
-
-const autoCompleteDriverFactory = inputWithOptionsDriverFactory;
-
-export default autoCompleteDriverFactory;
