+++ stories/Autocomplete/ExampleStandard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Autocomplete from 'wix-style-react/Autocomplete';
+import Autocomplete from 'wix-style-react/AutoComplete';
 import {array} from 'prop-types';
 
 const style = {
