+++ testkit/index.js
@@ -39,6 +39,9 @@ export const pageHeaderTestkitFactory = testkitFactoryCreator(pageHeaderDriverFa
 import languagePickerDriverFactory from '../src/LanguagePicker/LanguagePicker.driver';
 export const languagePickerTestkitFactory = testkitFactoryCreator(languagePickerDriverFactory);
 
+import autoCompleteDriverFactory from '../src/AutoComplete/AutoComplete.driver';
+export const autoCompleteTestkitFactory = testkitFactoryCreator(autoCompleteDriverFactory);
+
 import dropdownDriverFactory from '../src/Dropdown/Dropdown.driver';
 export const dropdownTestkitFactory = testkitFactoryCreator(dropdownDriverFactory);
 
@@ -201,7 +204,6 @@ export {
   badgeTestkitFactory,
   stylableCounterBadgeTestkitFactory as counterBadgeTestkitFactory,
   headingTestkitFactory,
-  autocompleteTestkitFactory,
   toggleSwitchTestkitFactory,
   textTestkitFactory
 } from 'wix-ui-backoffice/dist/src/testkit';
